+++ src/context/context-project-sub-category.ts
@@ -1,0 +1,6 @@
+import { createContext, Dispatch, SetStateAction } from "react";
+
+export const CategoryContext = createContext(0);
+export const SetCategoryContext = createContext<
+  Dispatch<SetStateAction<number>>
+>(() => {});
